+++ index.html
@@ -88,6 +88,13 @@
 
         <section id="works">
         
+            <div id="projectModal" class="projectModal">
+                <!-- Modal content -->
+                <div class="projectModal-content">
+                    <span class="close-projectModal">&times;</span>
+                    
+                </div>
+            </div>
 
         </section>
 
@@ -248,6 +255,6 @@
         </footer>
 
     </main>
-    <script src="./scripts.js"></script>
+    <script type="module" src="./scripts.js"></script>
 </body>
 </html>
