+++ style.css
@@ -451,6 +451,27 @@ textarea {
   cursor: pointer;
 }
 
+.projectModal {
+  display: none;
+  position: fixed;
+  z-index: 1;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  overflow: auto;
+  background-color: rgb(0, 0, 0); /* Fallback color */
+  background-color: rgba(0, 0, 0, 0.4); /* Black w/ opacity */
+}
+
+.projectModal-content {
+  background-color: #fefefe;
+  margin: 15% auto;
+  margin-top: 0;
+  padding: 20px;
+  height: 100%;
+}
+
 @media screen and (min-width: 992px) {
   #headline {
     height: 145vh;
@@ -528,6 +549,11 @@ textarea {
 
   .project-card .projectImage {
     flex: 1 0 45%;
+  }
+
+  .projectModal {
+    width: 60%;
+    height: 60%;
   }
 
   .aboutmeDescription {
